+++ client-rpc/src/server.rs
@@ -1,9 +1,9 @@
+use crate::program::Options;
 use crate::rpc::multisig_rpc::{MultiSigRpc, MultiSigRpcImpl};
 use crate::rpc::staking_rpc::{StakingRpc, StakingRpcImpl};
 use crate::rpc::sync_rpc::{SyncRpc, SyncRpcImpl};
 use crate::rpc::transaction_rpc::{TransactionRpc, TransactionRpcImpl};
 use crate::rpc::wallet_rpc::{WalletRpc, WalletRpcImpl};
-use crate::Options;
 use std::net::SocketAddr;
 
 use chain_core::init::network::{
